+++ tests/src/sanity.rs
@@ -72,7 +72,8 @@ mod tests {
 
     #[derive(Debug)]
     enum Application {
-        PythonHTTP,
+        PythonFlaskHTTP,
+        PythonFastApiHTTP,
         NodeHTTP,
         GoHTTP,
     }
@@ -105,6 +106,14 @@ mod tests {
 
         fn assert_stderr(&self) {
             assert!(self.stderr.lock().unwrap().is_empty());
+        }
+
+        fn assert_log_level(&self, stderr: bool, level: &str) {
+            if stderr {
+                assert!(!self.stderr.lock().unwrap().contains(level));
+            } else {
+                assert!(!self.stdout.lock().unwrap().contains(level));
+            }
         }
 
         fn assert_python_fileops_stderr(&self) {
@@ -179,13 +188,34 @@ mod tests {
             args: Option<Vec<&str>>,
         ) -> TestProcess {
             let process_cmd = match self {
-                Application::PythonHTTP => {
-                    vec!["python3", "-u", "python-e2e/app.py"]
+                Application::PythonFlaskHTTP => {
+                    vec!["python3", "-u", "python-e2e/app_flask.py"]
+                }
+                Application::PythonFastApiHTTP => {
+                    vec![
+                        "uvicorn",
+                        "--port=80",
+                        "--host=0.0.0.0",
+                        "--app-dir=./python-e2e/",
+                        "app_fastapi:app",
+                    ]
                 }
                 Application::NodeHTTP => vec!["node", "node-e2e/app.js"],
                 Application::GoHTTP => vec!["go-e2e/go-e2e"],
             };
             run(process_cmd, pod_name, namespace, args).await
+        }
+
+        fn assert(&self, process: &TestProcess) {
+            match self {
+                Application::PythonFastApiHTTP => {
+                    process.assert_log_level(true, "ERROR");
+                    process.assert_log_level(false, "ERROR");
+                    process.assert_log_level(true, "CRITICAL");
+                    process.assert_log_level(false, "CRITICAL");
+                }
+                _ => process.assert_stderr(),
+            }
         }
     }
 
@@ -555,7 +585,13 @@ mod tests {
         #[future]
         #[notrace]
         kube_client: Client,
-        #[values(Application::PythonHTTP, Application::NodeHTTP, Application::GoHTTP)] application: Application,
+        #[values(
+            Application::PythonFlaskHTTP,
+            Application::PythonFastApiHTTP,
+            Application::NodeHTTP,
+            Application::GoHTTP
+        )]
+        application: Application,
         #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
     ) {
         let service = service.await;
@@ -574,7 +610,8 @@ mod tests {
             .await
             .unwrap()
             .unwrap();
-        process.assert_stderr();
+
+        application.assert(&process);
     }
 
     #[cfg(target_os = "macos")]
@@ -589,7 +626,8 @@ mod tests {
         #[future]
         #[notrace]
         kube_client: Client,
-        #[values(Application::PythonHTTP)] application: Application,
+        #[values(Application::PythonFlaskHTTP, Application::PythonFastApiHTTP)]
+        application: Application,
         #[values(Agent::Job)] agent: Agent,
     ) {
         let service = service.await;
@@ -608,7 +646,8 @@ mod tests {
             .await
             .unwrap()
             .unwrap();
-        process.assert_stderr();
+
+        application.assert(&process);
     }
 
     #[cfg(target_os = "linux")]
@@ -783,7 +822,12 @@ mod tests {
     async fn test_steal_http_traffic(
         #[future] service: EchoService,
         #[future] kube_client: Client,
-        #[values(Application::PythonHTTP, Application::NodeHTTP)] application: Application,
+        #[values(
+            Application::PythonFlaskHTTP,
+            Application::PythonFastApiHTTP,
+            Application::NodeHTTP
+        )]
+        application: Application,
         #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
     ) {
         let service = service.await;
@@ -801,7 +845,8 @@ mod tests {
             .await
             .unwrap()
             .unwrap();
-        process.assert_stderr();
+
+        application.assert(&process);
     }
 
     #[rstest]